Clean up benchmark
