+++ packages/benchmarks/src/components/Wrapper/emotion.js
@@ -1,0 +1,14 @@
+import styled from 'react-emotion/macro'
+
+const Wrapper = styled('div')`
+  position: absolute;
+  transform-origin: 0 0;
+  left: 50%;
+  top: 50%;
+  width: 10px;
+  height: 10px;
+  background: #eee;
+  transform: scale(0.33);
+`
+
+export default Wrapper
