+++ packages/benchmarks/src/emotion.js
@@ -1,7 +1,11 @@
 import Box from './components/Box/emotion'
 import View from './components/View/emotion'
+import Dot from './components/Dot/emotion'
+import Wrapper from './components/Wrapper/emotion'
 
 export default {
   Box,
-  View
+  View,
+  Dot,
+  Wrapper
 }
